Fix CI caps and quality result edits
