+++ js/backup.js
@@ -7,13 +7,8 @@ import { profileStorageKey } from './profile-storage-key.js';
 import { getBlob, setBlob, shouldUseBlob } from './blob-storage.js';
 import { parseBackupSnapshot, serializeBackupSnapshot } from './backup-serialization.js';
 import { collectRawChatBackup } from './backup-chat-storage.js';
-import { clearProfileDeltaSnapshots } from './sync-delta-snapshot.js';
-import { markSyncProfileDirty } from './sync-dirty-state.js';
-import { markBackupRestorePending } from './sync-backup-restore-state.js';
-import {
-  getDailyRangeRaw,
-  upsertDailyBatchRaw,
-} from './wearables-store.js';
+import { prepareRestoredProfilesForSync } from './sync-backup-restore-state.js';
+import { getDailyRangeRaw, upsertDailyBatchRaw } from './wearables-store.js';
 import { VOICE_BACKUP_KEYS } from './voice-settings-schema.js';
 export { parseBackupSnapshot, serializeBackupSnapshot } from './backup-serialization.js';
 
@@ -122,33 +117,6 @@ async function writeRawStoredItem(key, value) {
   } else {
     localStorage.setItem(key, value);
   }
-}
-
-/**
- * A restored profile blob must win the first sync race after reload. Without
- * this reset, the per-row overlay can immediately reapply item tombstones
- * from the pre-restore state, making restored supplements briefly appear and
- * then disappear. Clearing planner snapshots also guarantees that an exact
- * restored copy is emitted again instead of being skipped by content hash.
- *
- * @param {any} backup
- */
-export function prepareRestoredProfilesForSync(backup) {
-  const profiles = Array.isArray(backup?.profiles) ? backup.profiles : [];
-  const prepared = new Set();
-  for (const profile of profiles) {
-    const profileId = profile?.profileId;
-    if (typeof profileId !== 'string' || !/^[a-zA-Z0-9_-]+$/.test(profileId) || prepared.has(profileId)) continue;
-    prepared.add(profileId);
-    clearProfileDeltaSnapshots(profileId);
-    // A restored blob may not have complete per-row history, so force the
-    // first recovery push to include the full v3 payload as a safety net.
-    try { localStorage.removeItem(`labcharts-${profileId}-sync-cutover-v2`); } catch {}
-    try { localStorage.removeItem(`labcharts-${profileId}-sync-ts`); } catch {}
-    markSyncProfileDirty(profileId);
-  }
-  markBackupRestorePending([...prepared]);
-  return prepared.size;
 }
 
 // ═══════════════════════════════════════════════
